requirements.txt
numpy>=1.21.0
librosa>=0.9.0
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.10.3
python-dotenv>=0.19.0
transformers>=4.30.0
scipy>=1.7.0
pydub>=0.25.1
streamlit>=1.24.0
openai>=1.0.0
